refactored accroding to PragDave principle. Added self healing.
